Test fix

+++ 0x07-pointers_arrays_strings/2-strchr.c
@@ -9,7 +9,7 @@
  */
 char *_strchr(char *s, char c)
 {
-	unsigned int i;
+	long unsigned int i;
 
 	for (i = 0; s[i] != '\0'; i++)
 	{
